[DC-2717] Rename to unshift
